requirements.txt
discord.py
asyncpg
aiocache
